[REFACTOR] 유저 도메인 리팩터링(뷰 합치기)

+++ gotbetter/src/main/java/pcrc/gotbetter/user/data_access/repository/UserRepository.java
@@ -7,6 +7,6 @@ import pcrc.gotbetter.user.data_access.domain.User;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer>, UserRepositoryQueryDSL {
+public interface UserRepository extends JpaRepository<User, Long>, UserRepositoryQueryDSL {
     Optional<User> findByAuthId(String auth_id);
 }
